requirements.txt
absl-py==0.2.0
altgraph==0.15
appnope==0.1.0
astor==0.6.2
attrs==17.3.0
biopython==1.71
bleach==1.5.0
bokeh==0.12.16
boto==2.48.0
boto3==1.4.8
botocore==1.8.11
bz2file==0.98
cachetools==2.1.0
cairocffi==0.9.0
CairoSVG==2.1.3
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
cssselect2==0.2.1
cycler==0.10.0
Cython==0.28.3
decorator==4.1.2
defusedxml==0.5.0
docopt==0.6.2
docutils==0.14
dominate==2.3.1
enum34==1.1.6
feather-format==0.4.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.4.0
flask-nav==0.6
Flask-Session==0.3.1
Flask-WTF==0.14.2
future==0.16.0
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.2
Keras==2.1.6
macholib==1.9
matplotlib==2.1.2
mysql-connector==2.1.4
mysqlclient==1.3.12
nltk==3.2.5
numpy==1.15.4
oauth2client==4.1.2
packaging==17.1
pandas==0.23.4
parso==0.1.1
pbr==4.1.0
pdfrw==0.4
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
psutil==5.4.6
ptyprocess==0.5.2
py==1.5.2
pyarrow==0.11.1
pycparser==2.18
Pydap==3.2.2
Pygments==2.2.0
PyInstaller==3.3.1
pyparsing==2.2.0
Pyphen==0.9.4
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
redis==2.10.6
requests==2.18.4
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.9.0
selenium==3.12.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.5.5
tensorflow==1.7.0
termcolor==1.1.0
tinycss2==0.6.1
tornado==4.5.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
weblogo==3.6.0
WebOb==1.8.2
Werkzeug==0.14.1
WTForms==2.1
